Use title attribute and split on \n
Instead of using the innerText and splitting on "|" or ".", we can use
the title tag of the span that surrounds the print goal and split on
"\n". This perhaps more accurately represents individual sprint goals
since we tend to line break after each goal. This also removes the need
to remember including either "|" or "." when writing our sprint goals as
an artificial separator.